+++ trieste/models/gpflow/models.py
@@ -14,8 +14,7 @@
 
 from __future__ import annotations
 
-from abc import abstractmethod
-from typing import Any, Optional, Protocol, Union
+from typing import Any, Optional, Union
 
 import gpflow
 import tensorflow as tf
@@ -38,15 +37,9 @@ from tensorflow_probability.python.util import TransformedVariable
 from ...data import Dataset
 from ...types import TensorType
 from ...utils import DEFAULTS, jit
-from ..interfaces import (
-    FastUpdateModel,
-    SupportsGetObservationNoise,
-    SupportsPredictJoint,
-    TrainableProbabilisticModel,
-    TrajectorySampler,
-)
+from ..interfaces import FastUpdateModel, TrainableProbabilisticModel, TrajectorySampler
 from ..optimizer import BatchOptimizer, Optimizer
-from .interface import GPflowPredictor
+from .interface import GPflowPredictor, SupportsCovarianceBetweenPoints
 from .sampler import RandomFourierFeatureTrajectorySampler
 from .utils import (
     assert_data_is_compatible,
@@ -54,37 +47,6 @@ from .utils import (
     randomize_hyperparameters,
     squeeze_hyperparameters,
 )
-
-
-class SupportsCovarianceBetweenPoints(SupportsPredictJoint, Protocol):
-    """A probabilistic model that supports covariance_between_points."""
-
-    @abstractmethod
-    def covariance_between_points(
-        self, query_points_1: TensorType, query_points_2: TensorType
-    ) -> TensorType:
-        r"""
-        Compute the posterior covariance between sets of query points.
-
-        .. math:: \Sigma_{12} = K_{12} - K_{x1}(K_{xx} + \sigma^2 I)^{-1}K_{x2}
-
-        Note that query_points_2 must be a rank 2 tensor, but query_points_1 can
-        have leading dimensions.
-
-        :param query_points_1: Set of query points with shape [..., N, D]
-        :param query_points_2: Sets of query points with shape [M, D]
-        :return: Covariance matrix between the sets of query points with shape [..., L, N, M]
-            (L being the number of latent GPs = number of output dimensions)
-        """
-        raise NotImplementedError
-
-
-class SupportsCovarianceObservationNoise(
-    SupportsCovarianceBetweenPoints, SupportsGetObservationNoise
-):
-    """A model that supports both covariance_between_points and get_observation_noise."""
-
-    pass
 
 
 class GaussianProcessRegression(
